apply API changes to BigDecimal that described in issue 1469

+++ beatrix/src/test/java/org/killbill/billing/beatrix/integration/TestIntegrationBase.java
@@ -1056,7 +1056,7 @@ public class TestIntegrationBase extends BeatrixTestSuiteWithEmbeddedDB implemen
                                    final BigDecimal amount,
                                    final CallContext context) throws UsageApiException {
         final List<UsageRecord> usageRecords = new ArrayList<>();
-        usageRecords.add(new UsageRecord(startDate, amount.longValue())); /* FIXME-1469 : API backward compat */
+        usageRecords.add(new UsageRecord(startDate, amount));
         final List<UnitUsageRecord> unitUsageRecords = new ArrayList<>();
         unitUsageRecords.add(new UnitUsageRecord(unitType, usageRecords));
         final SubscriptionUsageRecord record = new SubscriptionUsageRecord(subscriptionId, trackingId, unitUsageRecords);
